# docs/swagger.yaml
definitions:
  models.ApiResponse:
    properties:
      data: {}
      message:
        type: string
      success:
        type: boolean
    type: object
  models.User:
    properties:
      department:
        type: string
      email:
        type: string
      first_name:
        type: string
      id:
        type: integer
      last_name:
        type: string
      user_name:
        type: string
      user_status:
        type: string
    type: object
info:
  contact: {}
paths:
  /users:
    get:
      description: Retrieve a list of all users
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.User'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ApiResponse'
      summary: Get all users
      tags:
      - users
    post:
      consumes:
      - application/json
      description: Add a new user to the system
      parameters:
      - description: Create User
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.ApiResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ApiResponse'
      summary: Create a new user
      tags:
      - users
  /users/{id}:
    delete:
      description: Remove a user from the system by their ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ApiResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ApiResponse'
      summary: Delete a user
      tags:
      - users
    get:
      description: Retrieve a user by their ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ApiResponse'
      summary: Get a user by ID
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Update an existing user's details
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update User
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ApiResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ApiResponse'
      summary: Update a user
      tags:
      - users
swagger: "2.0"